Add new entity classes and repositories

+++ src/main/java/com/camayopolis/persistence/repository/IMovieRepository.java
@@ -3,5 +3,5 @@ package com.camayopolis.persistence.repository;
 import com.camayopolis.persistence.entity.MovieEntity;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface IMovieRepository extends ListCrudRepository<MovieEntity, Long> {
-}
+public interface IMovieRepository extends ListCrudRepository<MovieEntity, Integer> {
+  }
